Facebook registration and retrieval is now functional
Although it is funcitonal, it needs a lot of work as we need to
take that Facebook data and convert it to Interactions and then
we return the interactions.

Also updated the Schema and some model classes. Still need to clean-up the code a bit and really start working on the front end.

Signed-off-by: Vincent J. Palodichuk <vj.palodichuk@gmail.com>